Create Readme.md
12/01/2022
Local repository created for project and detailed study of already available program code is in progress.

18/01/2022
Tried running the project with some images of MNIST dataset of 0,1 and 2. Facing issue of exception  (System.InvalidOperationException) for variable 'tempStat'. Currently analysing it and studying abount HTM.

20/01/22
Test Run experiment with the MNIST dataset Images 0,1,2